MetaCTF/2024/RST/RE/grand-can-auto/client/requirements.txt
crccheck==1.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.8
packaging==24.1
pygame==2.6.0
Pygments==2.18.0
python-can==4.4.2
PyTMX==3.32
rich==13.8.0
typing_extensions==4.12.2
wrapt==1.16.0
